Clean ups: move defs to own file.
